+++ CHANGELOG.md
@@ -7,17 +7,28 @@
 
 ### Spring Core/Boot
 
+- Fix: Kotlin constructor with default values counts as autowired
+- Inspection: Warning: Spring @Value annotation string should start with "#{" "${" or resource prefixes
+
+### Other
+
+- License URL fix
+
+## [241.1.1167] - 2024-04-09
+
+### Spring Core/Boot
+
 - LineMarker: Navigation from Kotlin constructor parameters to property declaration
 
 ### Spring Data
 
 - Bean method doesn't navigate to autowired collection
 
-### Other 
+### Other
 
 - Upgrade to intellij 241 branch
 
-## [2024.233.1151] - 2024-04-08
+## 2024.233.1151 - 2024-04-08
 
 ### Spring Data